+++ security/src/main/scala/pl/touk/nussknacker/ui/security/oauth2/OAuth2Configuration.scala
@@ -43,6 +43,8 @@ case class OAuth2Configuration(usersFile: URI,
 
   override lazy val users: List[ConfigUser] = AuthenticationConfiguration.getUsers(userConfig).getOrElse(Nil)
 
+  def findUserById(id: String): Option[ConfigUser] = users.find(_.identity == id)
+
   def authorizeUrl: Option[URI] = Option(
     Uri(authorizeUri)
       .addParam("client_id", clientId)