+++ security/src/main/scala/pl/touk/nussknacker/ui/security/oauth2/OAuth2Profile.scala
@@ -1,10 +1,14 @@
 package pl.touk.nussknacker.ui.security.oauth2
 
-import pl.touk.nussknacker.ui.security.api.{AuthenticatedUser, AuthenticationConfiguration}
-import pl.touk.nussknacker.ui.security.api.AuthenticationConfiguration.ConfigUser
+import pl.touk.nussknacker.ui.security.api.AuthenticatedUser
+
+import scala.concurrent.{ExecutionContext, Future}
 
 trait OAuth2Profile[ProfileResponse] {
-  def getAuthenticatedUser(profile: ProfileResponse, configuration: OAuth2Configuration): AuthenticatedUser
+  def getAuthenticatedUser(accessTokenData: AccessTokenData,
+                           getProfile: => Future[ProfileResponse],
+                           configuration: OAuth2Configuration)
+                          (implicit ec: ExecutionContext): Future[AuthenticatedUser]
 }
 
 object OAuth2Profile {
@@ -17,7 +21,4 @@ object OAuth2Profile {
   def usernameBasedOnUsersConfiguration(identity: String, configuration: OAuth2Configuration): Option[String] =
     configuration.findUserById(identity).flatMap(_.username)
 
-  private implicit class RichOAuth2Configuration(c: OAuth2Configuration) {
-    private[OAuth2Profile] def findUserById(id: String) = c.users.find(_.identity == id)
-  }
 }