+++ security/src/test/scala/pl/touk/nussknacker/ui/security/oauth2/CachingOAuth2ServiceSpec.scala
@@ -40,6 +40,8 @@ class CachingOAuth2ServiceSpec extends AnyFunSpec with ScalaFutures with Matcher
       checkRecordings = checkRecordings + (accessToken -> (checkRecordings.getOrElse(accessToken, 0) + 1))
       jwtOAuth2Service.checkAuthorizationAndObtainUserinfo(accessToken)
     }
+    override def introspectAccessToken(accessToken: String): Future[AccessTokenData] = ???
+    override def obtainUserInfo(accessToken: String, accessTokenData: AccessTokenData): Future[OpenIdConnectUserInfo] = ???
   }
 
   private val cachingOAuth2Service = new CachingOAuth2Service(recordingJwtOauth2Service, config, ticker)