+++ security/src/test/scala/pl/touk/nussknacker/ui/security/oidc/GenericOidcTokenAuthenticationSpec.scala
@@ -1,4 +1,4 @@
-package pl.touk.nussknacker.ui.security.oauth2
+package pl.touk.nussknacker.ui.security.oidc
 
 import akka.http.javadsl.model.headers.HttpCredentials
 import akka.http.scaladsl.model.StatusCodes
@@ -16,54 +16,58 @@ import sttp.client3.testing.SttpBackendStub
 import sttp.model.{HeaderNames, Uri}
 
 import java.net.URI
-import java.security.KeyPairGenerator
 import java.time.Clock
-import java.util.Base64
 import scala.concurrent.Future
 import scala.concurrent.duration.DurationInt
 
-class JwtTokenAuthenticationSpec extends AnyFunSpec with Matchers with ScalatestRouteTest with Directives
+class GenericOidcTokenAuthenticationSpec extends AnyFunSpec with Matchers with ScalatestRouteTest with Directives
   with FailFastCirceSupport with OptionValues {
 
   implicit val clock: Clock = Clock.systemUTC()
 
-  private val keyPair = KeyPairGenerator.getInstance("RSA").generateKeyPair()
+  private val secret = "secret"
   private val userinfoUri = Uri(URI.create("http://authorization.server/userinfo"))
   private val audience = "http://nussknacker"
 
   implicit val timeout: RouteTestTimeout = RouteTestTimeout(5 seconds)
 
+  private val roleClaim = "http://namespace/roles"
+
   private val config = ConfigFactory.parseString(
     s"""authentication: {
-       |  method: "OAuth2"
+       |  method: "Oidc"
        |  usersFile: "classpath:oauth2-users.conf"
-       |  authorizeUri: "http://ignored"
-       |  clientSecret: "ignored"
+       |  issuer: "http://ignored"
+       |  authorizationEndpoint: "http://ignored"
+       |  tokenEndpoint: "http://ignored"
+       |  userinfoEndpoint: "$userinfoUri"
        |  clientId: "ignored"
-       |  profileUri: "$userinfoUri"
-       |  profileFormat: "oidc"
-       |  accessTokenUri: "http://authorization.server/token"
-       |  jwt: {
-       |    accessTokenIsJwt: true
-       |    publicKey: "${Base64.getEncoder.encodeToString(keyPair.getPublic.getEncoded)}"
-       |    audience: "$audience"
-       |  }
+       |  clientSecret: "$secret"
+       |  audience: "$audience"
+       |  rolesClaims: ["$roleClaim"]
        |}""".stripMargin)
 
-  private val validAccessToken = JwtCirce.encode(JwtClaim().about("admin").to(audience).expiresIn(180), keyPair.getPrivate, JwtAlgorithm.RS256)
-  private val expiredAccessToken = JwtCirce.encode(JwtClaim().about("admin").to(audience).expiresNow, keyPair.getPrivate, JwtAlgorithm.RS256)
-  private val accessTokenWithInvalidAudience = JwtCirce.encode(JwtClaim().about("admin").to("invalid").expiresIn(180), keyPair.getPrivate, JwtAlgorithm.RS256)
-  private val noProfileAccessToken = JwtCirce.encode(JwtClaim().about("no-profile-user").to(audience).expiresIn(180), keyPair.getPrivate, JwtAlgorithm.RS256)
+  private val validAccessToken = JwtCirce.encode(JwtClaim().about("admin").to(audience).expiresIn(180), secret, JwtAlgorithm.HS256)
+  private val expiredAccessToken = JwtCirce.encode(JwtClaim().about("admin").to(audience).expiresNow, secret, JwtAlgorithm.HS256)
+  private val accessTokenWithInvalidAudience = JwtCirce.encode(JwtClaim().about("admin").to("invalid").expiresIn(180), secret, JwtAlgorithm.HS256)
+  private val noProfileAccessToken = {
+    val claim = JwtClaim()
+      .about("no-profile-user")
+      .to(audience)
+      .withContent(
+        s"""{"$roleClaim": ["User"]}""".stripMargin)
+      .expiresIn(180)
+    JwtCirce.encode(claim, secret, JwtAlgorithm.HS256)
+  }
 
   implicit private val testingBackend: RecordingSttpBackend[Future, Any] = new RecordingSttpBackend(SttpBackendStub.asynchronousFuture
     .whenRequestMatches(req => req.uri == userinfoUri && req.header(HeaderNames.Authorization).value != s"Bearer $noProfileAccessToken")
     .thenRespond(s""" { "sub": "admin" } """))
-     // See classpath:oauth2-users.conf for the roles defined for user admin.
+  // See classpath:oauth2-users.conf for the roles defined for user admin.
 
   private val classLoader = getClass.getClassLoader
 
   private val authenticationResources = AuthenticationResources(config, classLoader, testingBackend)
-  assert(authenticationResources.isInstanceOf[OAuth2AuthenticationResources])
 
   private val testRoute = Route.seal(
     authenticationResources.authenticate() { authenticatedUser =>
@@ -98,7 +102,7 @@ class JwtTokenAuthenticationSpec extends AnyFunSpec with Matchers with Scalatest
   it("should not call for user profile when subject has username configured in user configuration") {
     Get("/config").addCredentials(HttpCredentials.createOAuth2BearerToken(noProfileAccessToken)) ~> testRoute ~> check {
       status shouldEqual StatusCodes.OK
-      responseAs[(String, Set[String])] shouldEqual ("SomeConfiguredUsername", Set("Admin"))
+      responseAs[(String, Set[String])] shouldEqual("SomeConfiguredUsername", Set("Admin", "User"))
     }
   }